#149: More refactoring on Windows matching with Linux
