+++ core-shared/webview/webview_c.h
@@ -1,0 +1,5 @@
+#pragma once
+
+extern "C" {
+	void web_view(const char* title, const char* url, int width, int height, int fullscreen, bool always_on_top, bool borderless, bool maximize, void *icon);
+}
