+++ core-windows/bin/app/assets/abc.json
@@ -1,0 +1,1 @@
+{f:5}
