+++ core-windows/bin/app/assets/abc.xml
@@ -1,0 +1,1 @@
+<A>test</A>
