+++ core-windows/build.bat
@@ -29,7 +29,7 @@ if EXIST bin\neutralino.exe (
     del /f bin\neutralino.exe 
 )
 
-g++  -std=gnu++17 -lstdc++fs -fconcepts -I ../core-shared ../core-shared/log.cpp src/main.cpp src/settings.cpp src/resources.cpp src/requestparser.cpp src/serverlistener.cpp src/functions.cpp src/core/computer/computer.cpp src/core/filesystem/filesystem.cpp src/core/os/os.cpp src/router.cpp src/auth/authbasic.cpp src/ping/ping.cpp src/core/storage/storage.cpp src/core/debug/debug.cpp src/core/app/app.cpp src/cloud/privileges.cpp src/webv.cpp -lws2_32 -static-libgcc -static-libstdc++ -Wl,-Bstatic -lstdc++ -lpthread -Wl,-Bdynamic -mwindows -Wconversion-null -DWEBVIEW_WINAPI=1 -lole32 -lcomctl32 -loleaut32 -luuid -lgdiplus -mwindows -lshlwapi -o bin/neutralino
+g++  -std=gnu++17 -lstdc++fs -fconcepts -I ../core-shared ../core-shared/log.cpp src/main.cpp ../core-shared/settings.cpp ../core-shared/resources.cpp src/server/requestparser.cpp src/server/serverlistener.cpp ../core-shared/helpers.cpp src/api/computer/computer.cpp src/api/filesystem/filesystem.cpp src/api/os/os.cpp ../core-shared/router.cpp ../core-shared/auth/authbasic.cpp ../core-shared/ping/ping.cpp src/api/storage/storage.cpp src/api/debug/debug.cpp src/api/app/app.cpp ../core-shared/cloud/privileges.cpp src/platform/windows.cpp ../core-shared/webview/webview_c.cpp -lws2_32 -static-libgcc -static-libstdc++ -Wl,-Bstatic -lstdc++ -lpthread -Wl,-Bdynamic -mwindows -Wconversion-null -DWEBVIEW_WINAPI=1 -lole32 -lcomctl32 -loleaut32 -luuid -lgdiplus -mwindows -lshlwapi -o bin/neutralino
 
 if EXIST bin\neutralino.exe (
     echo Neutralino binary is compiled in to bin/netralino.exe
