+++ core-windows/neutralinojs.log
@@ -1,0 +1,6 @@
+ERROR [src/settings.cpp:getSettings] std::exception
+ERROR [src/settings.cpp:getSettings] std::exception
+ERROR [src/settings.cpp:getSettings] std::exception
+ERROR [src/settings.cpp:getSettings] std::exception
+ERROR [src/settings.cpp:getSettings] std::exception
+ERROR [src/settings.cpp:getSettings] std::exception
