+++ core-windows/src/server/requestparser.cpp
