+++ core-windows/src/server/requestparser.h
