+++ core-windows/src/server/serverexceptions.h
